requirements.txt
fastapi==0.111.1
uvicorn==0.30.3

pydantic==2.8.2
sqlalchemy==2.0.32
openai==1.40.3
numpy==1.26.4
scikit-learn==1.5.0
tiktoken==0.7.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
ragas==0.1.15
datasets==2.21.0
nltk==3.9.1
scipy== 1.14.0
ragas==0.1.15
pandas==2.2.2
pymysql==1.1.1
black==24.8.0
authlib==1.3.2
python-jose==3.3.0
